+++ data/mypage/src/main/java/com/teamsolply/solply/mypage/source/MypageRemoteDataSource.kt
@@ -8,8 +8,8 @@ import com.teamsolply.solply.mypage.dto.response.PlaceTownListResponseDto
 interface MypageRemoteDataSource {
     suspend fun getPlaceTownList(): PlaceTownListResponseDto
     suspend fun getCourseTownList(): CourseTownListResponseDto
-    suspend fun getPlaceList(townId: Int): PlaceListResponseDto
-    suspend fun getCourseList(townId: Int): CourseListResponseDto
-    suspend fun deletePlaces(placeList: List<Int>)
-    suspend fun deleteCourses(courseList: List<Int>)
+    suspend fun getPlaceList(townId: Long): PlaceListResponseDto
+    suspend fun getCourseList(townId: Long): CourseListResponseDto
+    suspend fun deletePlaces(placeList: List<Long>)
+    suspend fun deleteCourses(courseList: List<Long>)
 }
