+++ domain/course/src/main/java/com/teamsolply/solply/course/model/CourseEntity.kt
@@ -3,7 +3,7 @@ package com.teamsolply.solply.course.model
 import com.teamsolply.solply.model.PlaceType
 
 data class CourseEntity(
-    val courseId: Int,
+    val courseId: Long,
     val courseName: String,
     val imageUrl: String,
     val tagList: List<PlaceType>,
