+++ domain/mypage/src/main/java/com/teamsolply/solply/mypage/model/CourseTownEntity.kt
@@ -3,7 +3,7 @@ package com.teamsolply.solply.mypage.model
 import com.teamsolply.solply.model.PlaceType
 
 data class CourseTownEntity(
-    val townId: Int,
+    val townId: Long,
     val townName: String,
     val courseName: String,
     val tagList: List<PlaceType>,
